Add files via upload
OhGodAnETHlargementPill is a tool that is said to increase the hash of the GTX 1080, GTX 1080 TI and Titan XP while mining Ethereum. EthlargementPill is a tool from OhGodACompany that improves Ethereum mining performance on high-performance NVIDIA-based GPUs with GDDR5X memory. It improves the hashrate not only on Ethereum, but also on all coins based on Ethash, and currently this tool only works with GTX 1080, GTX 1080TI and Nvidia TitanXP (GDDR5X memory). It is also compatible with both Windows and Linux, and you can find the release on GitHub.